Revert "Fully refactor firmware out of soccer"
This reverts commit 0f781acb226ec740c0688418d2744b3c47f6ee45.
