+++ soccer/radio/USBRadio.cpp
@@ -4,14 +4,15 @@
 #include <QMutexLocker>
 
 #include <Utils.hpp>
-#include "../common/firmware/common2015/drivers/cc1201/ti/defines.hpp"
+#include "../firmware/common2015/drivers/cc1201/ti/defines.hpp"
 #include "USBRadio.hpp"
 
 // Include this file for base station usb vendor/product ids
-#include "../common/firmware/base2015/usb-interface.hpp"
+#include "../firmware/base2015/usb-interface.hpp"
 
 // included for kicer status enum
-#include "../common/firmware/robot2015/cpu/status.h"
+#include "../firmware/robot2011/cpu/status.h"
+
 
 using namespace std;
 using namespace Packet;
